+++ src/common/components/auth/stellar/ImportAccount.tsx
@@ -14,7 +14,7 @@ import {
 } from '../../ui/dialog';
 import { Input } from '../../ui/input';
 
-import useStellar from '@/services/stellar/hook/useStellar';
+import useStellar from '@/modules/stellar/hook/useStellar';
 
 enum SECRET_KEY_ERROR {
 	REQUIRED = 'Secret key is required',
